Batch multiple data changes into a single observer call
Some semantic forms can trigger a lot of synchronous data changes and as a consequence a lot of observer callbacks. By batching these data changes together we can reduce the amount of callbacks which potentially improves performance in case the handler do heavy work.

It also makes it possible to analyse the data and remove redundant changes so the observers are only triggered if there are actual changes.